Authentication added for JournalEntry points

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -10,6 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Component
@@ -22,36 +23,42 @@ public class JournalEntryService {
     private UserEntryService userEntryService;
 
     @Transactional
-    public void saveEntry(JournalEntry journalEntry, String userName)
-    {
+    public void saveEntry(JournalEntry journalEntry, String userName) {
         User user = userEntryService.findByUserName(userName);
         journalEntry.setDate(LocalDateTime.now());
         journalEntryRepository.save(journalEntry);
         user.getJournalEntries().add(journalEntry);
-        userEntryService.saveEntry(user);
+        userEntryService.saveUser(user);
     }
 
-    public void saveEntry(JournalEntry journalEntry)
-    {
+    public void saveEntry(JournalEntry journalEntry) {
         journalEntry.setDate(LocalDateTime.now());
         journalEntryRepository.save(journalEntry);
     }
 
-    public List<JournalEntry> getAll()
-    {
+    public List<JournalEntry> getAll() {
         return journalEntryRepository.findAll();
     }
 
-    public Optional<JournalEntry> findById(ObjectId id)
-    {
+    public Optional<JournalEntry> findById(ObjectId id) {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(String userName, ObjectId id)
-    {
-        User user  = userEntryService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        journalEntryRepository.deleteById(id);
-        userEntryService.saveEntry(user);
+    @Transactional
+    public boolean deleteById(String userName, ObjectId id) {
+        boolean removed = false;
+        try {
+            User user = userEntryService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed){
+                journalEntryRepository.deleteById(id);
+                userEntryService.saveUser(user);
+            }
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the entry", e);
+        }
+
+        return removed;
     }
 }
